chore: sanity check notebooks are functional (#18)
* chore: went through notebooks once to verify they still work

* chore: added directory with core notebooks to be run in pull requests as additional tests

* feat: added run-notebooks.yml workflow to execute core notebooks

* chore: updated paths to store models

* fix: replace python with uv setup step

* fix: switch from nbconvert to execute

+++ src/random_neural_net_models/data.py
@@ -94,7 +94,7 @@ class NumpyNumCatTrainingDataset(Dataset):
                     for i, v in enumerate(x_cat)
                 ]
             ],
-            dtype=np.integer,
+            dtype=np.int64,
         )
 
     def __getitem__(self, idx: int) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
@@ -152,7 +152,7 @@ class NumpyNumCatTrainingDatasetXOnly(Dataset):
                     for i, v in enumerate(x_cat)
                 ]
             ],
-            dtype=np.integer,
+            dtype=np.int64,
         )
 
     def __getitem__(self, idx: int) -> tuple[torch.Tensor, torch.Tensor]:
